databricks_app/requirements.txt
streamlit>=1.35.0
databricks-sql-connector>=3.4.0
databricks-sdk>=0.28.0
pydantic>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
